lexer/lexer.go
package lexer

import "monkey/token"

type Lexer struct {
	input       string
	currentChar byte
	currentPos  int
	nextPos     int
}

func New(input string) *Lexer {
	l := &Lexer{input: input}

	if len(input) > 0 {
		l.currentChar = input[0]
		l.nextPos = 1
		// no need to set l.currentPos to 0, Go already instanciates it with its struct type's zero value
	}

	return l
}

func (l *Lexer) readChar() {
	if l.nextPos >= len(l.input) {
		l.currentChar = 0
	} else {
		l.currentChar = l.input[l.nextPos]
	}
	l.currentPos = l.nextPos
	l.nextPos++
}

func (l *Lexer) NextToken() token.Token {
	var token token.Token
	return token
}
